+++ src/OpenTelemetry.Instrumentation.EventCounters/CHANGELOG.md
@@ -3,8 +3,8 @@
 ## Unreleased
 
 * Update OpenTelemetry.Api to 1.3.1.
-* Changed `EventCounter` prefix to `ec` and started abbreviating event source name
-  when instrument name is longer that 63 characters.
+* Change `EventCounter` prefix to `ec` and abbreviate event source name
+  when instrument name is longer than 63 characters.
   ([#740](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/740))
 
 ## 1.0.0-alpha.1